attempted equivalent exchange
particles still duplicate and disappear when they aren't supposed to.
equivalent exchange is very important to me in this project, and
although the duplication and disappearing aren't as bad in this patch,
they are still there. I will continue to experiment, or I may try to
find a way to make the performance of the 'objects' branch acceptable.
(that branch treats all cells as instances which can vary in number,
rather than viewing the size of the board as a constant amount of
cells that can be in different states.)